[2/4] sqlite query perf: try to reuse in-flight statements
If we avoid resetting statements, we may be able to continue where we
left off last time. This saves us the work of repeating the whole query
over and over again when we just want to fill sequential windows.

However, leaving statements open like that ("busy", in sqlite terms)
may cause some queries to return SQLITE_LOCKED (example: DROP TABLE).
So, we have to add some logic to reset any busy statements whenever
a query fails with that error. This meant that we had to be able to
call up to Java from native, which forced me to change a few methods
from static to instance, which makes the patch look bigger than it
really is.

Leaving statements open also keeps sqlite's internal query state,
increasing memory usage. We'll mitigate this in the next patch.

My benchmark - moveToNext() through a 25000-row result set, with the
window being able to fit around 6700 - improved from 1350ms to 750ms.

Change-Id: I03af49d8b1bade3e63c00f3fe5ef3259989c292c
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: spezi77 <spezi77@gmx.com>