Fix window transformation related issue
In some cases, for example, a window containing a SurfaceView,
when pressing switch button, the window animation of SurfaceView
and it's attached window are not sync, which leads animation not
behaving correctly.

To avoid this, apply the translation that applies the position
of the window before the app transformation.

Google issue:
https://code.google.com/p/android/issues/detail?id=225134

Change-Id: Id1a1b3ffc8bc553e55f5512f3ebabb5b154516cd
Signed-off-by: spezi77 <spezi7713@gmx.net>