Vectorized the power menu sound panel selector
Reason for doing this is because it allows us to use the AOSP themes that are now
built-in and not have a lot of jank. The power menu selector will adapt to the
accent and give us a nice look :-)

This will also come in handy for Substratum themes in where the themer either forgets
about the selector and so keeps things looking less janky!

Change-Id: I614406862bfe92d61f64c4924c1d6c3cc2d79570

Signed-off-by: spezi77 <spezi77@gmx.com>