Always send configuration if the boundaries change
The dialog widths can be wrong on orientation change if maxAspectRatio
is declared for an activity. That risk is eliminated by always sending
the new Configuration to the activity if the calculated boundaries will
lead to an update of the MergedConfiguration.

Test: Install Use android:maxAspectRatio=1.5 in the manifest on a
PreferenceActivity that opens an input dialog on a pixel device.
Rotate the device and you can get the wrong width of the dialog.

Bug: 70326569
Test: Running test application in associated bug.
Change-Id: I8d0caff80c64103df75091b19ee736e7f46ca7d9
Signed-off-by: mydongistiny <jaysonedson@gmail.com>