base: add settings based hook for default outgoing voice account changed
yes this is a little hack to tame qti telephony blobs
but it does not hurt anyone aside from not beeing pride about it
and I hope we can get rid of that at some point

since its not possible to listen for all possible changes here
e.g. ask every time is not changing the value of MULTI_SIM_VOICE_CALL_SUBSCRIPTION
add this little hook that allows client to listen to every change
to default outgoing phone account from undefined to defined

this allows client to listen with a ContentObserver to get notified
whenever anything changes.

Change-Id: Ifeea1f16f88a689011db63f5b5ecba1a09701dff
Signed-off-by: spezi77 <spezi77@gmx.com>