SystemUI: Port Now Playing feature from Pixel [1/2]
All of the code here was reverse engineered from smali and a
decompiled version of SystemUIGoogle from the taimen factory
images.

The decompiler wasn't a fan of the Lambdas in the code (probably
beacause they were previously obfuscated in a dex) so I converted
them into separate classes where needed.

The other commit adds the necessary files for music recognition.

Change-Id: Iac7b4c18b746b4048eb8fd8037f5b54b0eef36c6
Signed-off-by: Joe Maples <joe@frap129.org>
Signed-off-by: spezi77 <spezi7713@gmx.net>