Hold back key to force kill [1/2] this is based on Danesh works from : Change-Id: Ia88d8a2753e203afb8d6117a643c33127f991c4a
However, it is modified to not use the TaskUtil class from Omni.
It also kill only the first foreground task and not background task
like GSM. It's also hooked up to the original long press event
instead of making a new event listener.

Finally, it will also tell which app was killed in the toast.

Signed-off-by: spezi77 <spezi7713@gmx.net>