Introduce Ambient Indication Interface
The is a interface that handles subtle indications in Ambient Display
and the Lockscreen. To start, we've rewritten our Ambient Play feature
from scratch to be housed in this interface.

Ambient Play:

A music recognition service that records data from the users surroundings,
to then submit that data to a fingerprint matching database. If a match
is found, it returns a result. That result is then parsed and display via
the AmbientIndicationContainer for Ambient Display and Lockscreen, as well
as a non-intrusive notification.

Some of the improvements we've made from our previous iteration is battery
and performance. Ambient Play now listens for data every 120 seconds instead
of 60 seconds. Ambient Play also auto disables when a device is charging.
This will prevent possible slow charging and rare discharging due to the
amount of background proccessing.

In our previous iteration, the recording thread wasn't stopping properly
and would almost always run continuosly, even when the user disabled the
feature. We've added an extensive ammount of calls to stop the thread after
multiple events, to insure the thread only runs to serve it's one purpose
then goes away.

test: Battery and performance is not affected as much as before.

Ambient Container:

Developed by Google and backported by frap129 although we're using a
slimmed down version of the container from ezio84

todo:

- Add double tap events, to submit data to Google Search
- Add support for displaying history from matched results in Settings
- Add Ambient Weather

Change-Id: Iea2c70f17fa5d9aee48bd2efe1fa863b2cbd4308

PixelDust Edits:
Small alignements to work together with  ambient music ticker code
written by ezio.

Signed-off-by: spezi77 <spezi7713@gmx.net>