Aosp Recents: bigger icons, dynamic mediaplaying and alarm clock infos
If a media player is playing, the related Recents card will get the
albumart as app icon with a fade effect, the album color as background
color, track duration and title as app title.

If the app is Clock and an alarm is active, alarm info are appended to
the card app name.

Also use bigger app icons (we have iconpacks support, let's taste it
better) and allow two lines app title.

Giving some Slim Recents magic to stock Recents :D

Change-Id: Ie38f5f222f77a1eb791c28f5ee14629b7be6794a
Signed-off-by: spezi77 <spezi7713@gmx.net>